--- test/fixture/gyp/binding.gyp
@@ -1,8 +1,0 @@
-{
-  "targets": [
-    {
-      "target_name": "hello",
-      "sources": [ "hello.cc" ]
-    }
-  ]
-}
