--- test/fixture/gyp/index.html
@@ -1,22 +1,0 @@
-<!doctype html>
-
-<html>
-  <head>
-    <title>NW.js Node Native Addon Demo</title>
-    <script>
-      window.onload = () => {
-        const { hello } = require("./build/Release/hello");
-        document.getElementById("test").textContent = hello();
-      };
-    </script>
-  </head>
-
-  <body>
-    <span id="test"></span>
-    <br />
-    <span
-      >If you can see the "world" text above, then the Node Native module was
-      successful</span
-    >
-  </body>
-</html>
