--- test/fixture/gyp/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "gyp",
-  "main": "index.html"
-}
