+++ test/specs/get.test.js
@@ -2,9 +2,10 @@ import assert from "node:assert";
 import fs from "node:fs";
 import path from "node:path";
 import process from "node:process";
-import { describe, it, before } from "node:test";
 
-import get from '../src/get.js';
+import { beforeAll, describe, it } from "vitest";
+
+import get from '../../src/get.js';
 
 describe("get", async () => {
   const nwOptions = {
@@ -19,7 +20,7 @@ describe("get", async () => {
     nativeAddon: false,
   };
     
-  before(async () => {
+  beforeAll(async () => {
     await get(nwOptions);
   });
